// ppll-native-client/src/renderer/src/types/binance.ts
export interface AccountPosition {
  symbol: string
  initialMargin: string
  maintMargin: string
  unrealizedProfit: string
  positionInitialMargin: string
  openOrderInitialMargin: string
  leverage: string
  isolated: boolean
  entryPrice: string
  breakEvenPrice: string
  liquidationPrice: string
  maxNotional: string
  positionSide: 'BOTH' | 'LONG' | 'SHORT'
  positionAmt: string
  notional: string
  isolatedWallet: string
  updateTime: number
  bidNotional: string
  askNotional: string
}